Handle block shapes up to 7x7

+++ packets/src/structures/installed_block.rs
@@ -8,7 +8,7 @@ pub struct InstalledBlock {
     pub variant: usize,
     pub rotation: u8,
     pub color: BlockColor,
-    pub position: (usize, usize),
+    pub position: (i8, i8),
 }
 
 impl InstalledBlock {
